// src/types.d.ts
export interface LINEMusicOptions {
    lmlc: string;
    deviceId: string;
}
export declare enum SortType {
    Relevance = "RELEVANCE",
    Popular = "POPLAR",
    Release = "RELEASE"
}
export declare enum SearchType {
    Tracks = "tracks",
    Albums = "albums",
    Artists = "artists",
    Videos = "videos",
    Playlists = "playlists",
    Lyrics = "lyrics"
}
export declare enum FeaturedType {
    Tracks = "tracks",
    Albums = "albums"
}
export declare enum ChartType {
    Tracks = "tracks",
    Albums = "albums",
    Videos = "videos"
}
export declare enum BitRateType {
    Min = "min",
    Mid = "mid",
    Max = "max"
}
export interface BaseResponse<T> {
    response: {
        result: T;
    };
}
export interface SearchOptions {
    start: number;
    display: number;
    sortType: SortType;
    searchType: SearchType;
}
export interface DisplayOptions {
    start: number;
    display: number;
}
export interface LyricsOptions {
    apiVersion: number;
    nonSync: boolean;
}
export interface GetTrackSourceOptions {
    bitRateType: BitRateType;
}
// ...
export interface Track {
    trackId: string;
    trackTitle: string;
    discNumber: number;
    trackNumber: number;
    artistTotalCount: number;
    listenedCount: number;
    artists: Artist[];
    album: Album;
    hasLyric: boolean;
    isStreaming: boolean;
    isDownload: boolean;
    isMobileDownload: boolean;
    userAction: {
        isPurchased: boolean;
    };
    likeCount: number;
    isKaraokeEnabled: boolean;
    isExplicit: boolean;
    lyric?: string;
}
export interface Artist {
    artistId: string;
    artistName: string;
    trackCount: number;
    albumCount: number;
    videoCount: number;
    imageUrl: string;
    likeCount: number;
}
export interface Album {
    albumId: string;
    albumTitle: string;
    releaseDate: string;
    imageUrl: string;
    artistTotalCount: number;
    artists: Artist[];
    trackTotalCount: number;
    isAdult: boolean;
    producerLine: string;
    isStreaming: boolean;
    isMobileDownload: boolean;
    isDownload: boolean;
    likeCount: number;
    userAction: {
        isPurchased: boolean;
    };
}
export interface Video {
    videoId: string;
    imageUrl: string;
    videoTitle: string;
    artistName: string;
    playTime: string;
    releaseDate: string;
    track: Track;
}
export interface Playlist {
    plId: string;
    title: string;
    subTitle: string;
    image: {
        baseImageUrl: string;
    };
    desc: string;
    timeInfo: string;
    updateTime: string;
    tracksTotalCount: number;
    sub2Title: string;
    playtime: number;
    likeCount: number;
    isPublic: boolean;
    isMine: boolean;
    type: string;
    contentType: string;
    playlistContentType: string;
    playlistDataType: string;
    isCoverImageUrl: boolean;
    isOfficialAccount: boolean;
}